"N-Queens Visualizer" - Exploring Chessboard
The "N-Queens Visualizer" is a captivating web application that offers an engaging solution to one of the most intriguing puzzles in the world of chess and computer science – the N-Queens problem. In this classic puzzle, the challenge is to place N chess queens on an NxN chessboard in such a way that no two queens threaten each other. While the problem itself is mathematically fascinating, our project takes it a step further by providing an interactive platform to visualize and explore these solutions.

Key Features

Interactive Input: The core of this web application revolves around its user-friendly interface. Users can input the number of queens (N) they want to explore solutions for, making it accessible to chess enthusiasts, students, and curious minds alike.

Dynamic Chessboard Rendering: Our application leverages the power of HTML, CSS, and JavaScript, coupled with the D3.js library, to create a visually captivating chessboard. Queens are placed dynamically on the board according to the constraints of the N-Queens problem. It offers real-time visualization of solutions, making learning and exploration an interactive experience.

Random Solution Generation: To enhance the user experience, the "N-Queens Visualizer" doesn't stop at a single solution. It randomly selects and displays solutions, providing a diverse range of configurations to study. This feature encourages deeper exploration and understanding of the problem's nuances.

User-Friendly Design: We understand the importance of a seamless user experience. Hence, the application is designed to be highly interactive and user-friendly. Clear error handling ensures that users input a valid value for N (with a minimum of 4), and informative feedback is provided in case of invalid inputs.

Why the N-Queens Problem Matters

Beyond its status as a captivating puzzle, the N-Queens problem holds significant importance in computer science, artificial intelligence, and optimization. Its solutions have practical applications in diverse fields such as scheduling, layout design, and DNA sequence analysis. This project provides a practical and engaging way to explore the fundamental concepts underlying these applications.

How to Engage with the "N-Queens Visualizer"

Input the number of queens (N) you want to explore and click the "Submit" button.
Dive into the world of chessboard configurations, where queens are strategically placed to meet the problem's constraints.
Experiment with different N values and observe how solutions vary.
Reset the board at any time to embark on a new exploration.
In Conclusion

The "N-Queens Visualizer" project isn't just a puzzle-solving tool; it's an educational and eye-opening experience. Whether you're a chess enthusiast, a student of computer science, or simply